combine static announcements into single component
