--- packages/blockchain-wallet-v4-frontend/src/components/Announcement/Static/EmailReminder/index.js
@@ -1,36 +1,0 @@
-import React from 'react'
-import { connect } from 'react-redux'
-import { bindActionCreators } from 'redux'
-
-import { actions } from 'data'
-import EmailReminder from './template'
-
-class EmailReminderContainer extends React.PureComponent {
-  handleResend = email => {
-    if (this.state.reminded) return
-    this.props.actions.resendVerifyEmail(email)
-    this.setState({ reminded: true })
-    setTimeout(() => {
-      this.setState({ reminded: false })
-    }, 3000)
-  }
-
-  render () {
-    return (
-      <EmailReminder
-        email={this.props.email}
-        reminded={this.state.reminded}
-        handleResend={this.handleResend}
-      />
-    )
-  }
-}
-
-const mapDispatchToProps = dispatch => ({
-  actions: bindActionCreators(actions.modules.securityCenter, dispatch)
-})
-
-export default connect(
-  null,
-  mapDispatchToProps
-)(EmailReminderContainer)
